generics, refactor
