+++ chapter6/src/test/java/com/backstopmedia/kotlin/chapter6/collections/Arrays.kt
@@ -1,4 +1,4 @@
-package com.backstopmedia.kotlin.chapter6
+package com.backstopmedia.kotlin.chapter6.collections
 
 import org.junit.Test
